remove useles layout
